Migrate setup.py and setup.cfg to pyproject.toml
Move the custom sdist command from wagtail/utils/setup.py directly to
setup.py, because we cannot import the wagtail package itself during
the build process without installing it in the build environment.
