--- setup.cfg
@@ -1,11 +1,0 @@
-[bdist_wheel]
-python_tag = py3
-
-[doc8]
-ignore = D001
-ignore_path = _build,docs/_build
-
-[tool:pytest]
-django_find_project = false
-python_files=test_*.py
-testpaths=wagtail
